requirements.txt
# requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pymongo[srv]
voyageai
ollama
python-multipart
jupyterlab